% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LICOR_processing.R
\name{RAW_file_cleanup}
\alias{RAW_file_cleanup}
\title{Clean up a raw LI-COR file for data analysis}
\usage{
RAW_file_cleanup(dat)
}
\arguments{
\item{dat}{A datafile called "dat" --> the raw LI-COR data loaded as a CSV}
}
\value{
A cleaned up version of the data
}
\description{
Removes all the extra data at the top of CSV regarding LI-COR conditions.
Moves the column names from the middle of the dataset to the top.
}
